Revert "Development davidson march20"
