+++ functions/package.json
@@ -30,8 +30,7 @@
     "firebase-tools": "^3.17.4",
     "morgan": "~1.9.0",
     "selenium-webdriver": "^3.6.0",
-    "sinon": "^4.4.2",
-    "three-js": "^79.0.0"
+    "sinon": "^4.4.2"
   },
   "devDependencies": {
     "chai": "^4.1.2",
@@ -40,7 +39,6 @@
     "firebase-server": "^0.12.0",
     "geckodriver": "^1.10.0",
     "mocha": "^5.0.0",
-    "three": "^0.90.0",
     "nyc": "^11.6.0"
   }
 }
