+++ functions/views/VRWorld.ejs
@@ -23,7 +23,7 @@
   <script src="/javascripts/auth.js"></script>
   <script src="/javascripts/newWorld.js"></script>
   <script src="/javascripts/pointSelection.js"></script>
-  <script src="/javascripts/controls.js"></script>
+  <script src="/javascripts/pointStatistics.js"></script>
   <script src="https://cdn.rawgit.com/dataarts/dat.guiVR/master/build/datguivr.js"></script>
 
   <script type="x-shader/x-vertex" id="vertexshader">
